streamlit_app/requirements.txt
streamlit==1.14.0
streamlit-authenticator==0.2.1
pandas==1.5.1
plotly==5.11.0
boto3==1.26.12
PyYAML==6.0
